rif: fix type of cast target
- fixes b9c22795e2842cd52f4cb495ab9f7544976f5e33

+++ src/rif/gspc-bootstrap-rifd.cpp
@@ -110,7 +110,7 @@ try
               )
           , vm.count (option::port)
           ? boost::make_optional<unsigned short>
-            ( static_cast<unsigned long>
+            ( static_cast<unsigned short>
               ( vm.at (option::port)
               . as<fhg::util::boost::program_options::positive_integral<unsigned short>>()
               )
